Move appropriate tests to cli tests and change setup entry point.
